get_new_records/requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.14
h11==0.13.0
idna==3.3
outcome==1.1.0
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.27.1
selenium==4.1.1
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.10.0
urllib3==1.26.8
wrapt==1.16.0
wsproto==1.0.0
